requirements.txt
python-docx
ukrainian-word-stress